type inconsistency resolved

+++ test_GMM_mpi.cpp
@@ -79,8 +79,8 @@ int main(int argc, char ** argv)
         int _simulation_length = SIMULATION_LENGTH;
         double _c_factor = C;
         double _mh_target_acc = MH_TARGET_ACC;
-        double _energy_level_tuning_max_time = ENERGY_LEVEL_TUNING_MAX_TIME;
-	double highest_level = -1; 
+        int _energy_level_tuning_max_time = ENERGY_LEVEL_TUNING_MAX_TIME;
+	int highest_level = -1; 
 
 	/* parse command line options */
 	int opt;
